Factor out add_sub code
